finished... for now

+++ img/entrypoint.py
@@ -53,9 +53,10 @@ if not os.path.exists('/minecraft/config/whitelist.json'):
     fp = open('/minecraft/config/whitelist.json', 'w')
     fp.close()
 
-os.symlink('/minecraft/config/usercache.json', '/minecraft/usercache.json')
-os.symlink('/minecraft/config/whitelist.json', '/minecraft/whitelist.json')
-os.symlink('/minecraft/config/ops.json', '/minecraft/ops.json')
-os.system('"echo \\"$(cat /minecraft/server.properties)\\""> /minecraft/server.properties')
+if not os.path.exists('/minecraft/usercache.json'):
+    os.symlink('/minecraft/config/usercache.json', '/minecraft/usercache.json')
+    os.symlink('/minecraft/config/whitelist.json', '/minecraft/whitelist.json')
+    os.symlink('/minecraft/config/ops.json', '/minecraft/ops.json')
+    os.system('"echo \\"$(cat /minecraft/server.properties)\\""> /minecraft/server.properties')
 
 os.system('java -cp /minecraft -Xmx${JAVA_MEMORY} -Xms${JAVA_MEMORY} -Dfml.queryResult=confirm -jar forge-*.jar nogui')
